pyproject.toml
[tool.poetry]
name = "prono"
version = "0.1.0"
description = "Project Management with Real-Time Collaboration (ASGI)"
authors = ["Omar Yesith Alvarado González <omaryesith@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
django-ninja = "^1.0"
psycopg = "^3.1"
celery = "^5.3"
python-decouple = "^3.8"
# ASGI / Real-Time
channels = "^4.0"
channels-redis = "^4.2"
daphne = "^4.1"
# Security
django-ninja-extra = "*"
django-ninja-jwt = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
black = "^24.1"
isort = "^5.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
